+++ Samba Action Calculator/SACBlockAction.h
@@ -17,7 +17,7 @@
 - (id)initWithValues:(NSInteger)at
              minRoll:(NSInteger)mn
            forPlayer:(SACPlayer *)player
-             atIndex:(int)index
+             atIndex:(NSInteger)index
         successArr:(NSArray*)sa;
 
 - (NSMutableString*) getDiceDescription;
